Opcode arg work: cleanup and changes to facilitate shapes
Instead of passing the individual members of an OpcodeArgValue to the typeinfo functors, we now just hass the whole OpcodeArgValue instance. Moreover, the functors for postprocessing and converting to a string now return the number of bits they have read, which is necessary to handle cases where one type contains multiple instances of another type (e.g. a shape with multiple number variables as its dimensions: it needs to know where one number variable has ended so it knows where to start the next one off).

Finished the "shape" opcode arg type.

Streamlined the generic system for opcode arg variables slightly: removed the ability to specify custom functors for decompiling and stringifying, since no variable types used it.