Opcode arg work: backporting generic variable system
I want to backport the generic variable handling to the old opcode argument system. If I get that working *and* remove the need for postprocessing, then I basically have no reason to even continue with the new opcode arg system.

Started work on this, by creating a generic Variable class and backporting the scope-indicator stuff. I can subclass that Variable class for the typed variables.

Made several fixes needed so that the program could compile, and addressed compiler warnings as well.