Opcode arg work: backporting generic variable handling
The old opcode argument system now handles number, object, and player variables in a generic fashion.

Some band-aid fixes were needed to the read and write methods on Condition and Action in order to let us pass typeinfos; if we try to pass TYPE::typeinfo.factory to an opcode function definition directly, we end up passing nullptr because the opcode functions are actually constructed before the typeinfo instances (because static variables in this language are a nightmare); as such, in order to test just the new stuff I needed some band-aids so I could construct an argument definition with a factory function OR a typeinfo reference.