Player-or-group type; critical fix to resaving references to script stats
Fixed a critical bug that prevented us from resaving references to scripted stats properly; it appeared sometime after we introduced cobb::indexed_list (after the last release, thank goodness) as a result of ReachMegaloGameStat having its own (index) member, when it should've only had the one provided by its superclass, cobb::indexed_refcountable.

Added a "player or group" type to replace the OpcodeArgTeamOrPlayerVariableFactory factory. By using a separate type, we gain the ability to have the type itself serialize the team-or-player enum, meaning that we no longer need to check for the factory in the opcode save functions.