Enum argument types, except loadout palettes
Moved DetailedEnum and DetailedFlags to the "formats" folder.

All enum types that exist in the old system have been upgraded to use DetailedEnum. We still need to convert loadout palette indices to an enum type.

Fixed opcode-to-script mappings for Modify Score. Fixed decompiling for property-setters where the property to be set is identified by an argument (in practice, this is only Modify Grenade Count).