Decompiler progress
OpcodeArgValue::decompile now uses a common flags type, Decompiler::flags_t. A flag is now used to signal the context of a thiscall, allowing OpcodeArgValueObjectTimerVariable to decompile in a manner that allows it to be both the context of a call and the argument of a call.

Removed the unnecessary superclass for OpcodeArgValueObjectTimerVariable.

All argument types except formatted strings can now be decompiled. All argument types now define a typeinfo struct, and all opcode function definitions now refer to that struct; directly referring to a factory function is no longer possible, and OpcodeArgBase has been modified appropriately.

Deleted OpcodeArgValueBaseEnum, OpcodeArgValueBaseFlags, and OpcodeArgValueAllPlayers, as they are now unused.