Triggers no longer postprocess for label loops
A trigger of the form (for each object with label "foo") no longer needs to rely on postprocessing to obtain a pointer to the target Forge label.

Next task: I need to write code to "flatten" a loaded trigger hierarchy into flat lists of conditions, actions, and triggers, even when conditions and actions are interleaved together. (Not sure why Bungie used flat lists by type. They could've interleaved opcodes, prefixing each with a bit to indicate the type; the only way the current approach would be more efficient (given the need for 10-bit action indices on conditions to allow interleaving) is if most scripts have at least ten times as many conditions as actions... which I suppose I should check.)