Trigger flattening for serialization is DONE
Trigger serialization is done. My initial approach was almost perfect right out of the gate; I just forgot to add a flag to triggers to indicate that they've already been flattened (which is needed since we flatten nested triggers first, but iterate over the whole trigger list in the first place). As a bonus, that flag also allows us to have triggers that are referenced from multiple places int the vein of subroutines (though it's not yet clear whether the engine supports this).

No major tasks remain. I can begin work on the compiler at any time.