Work on MPVR space utilization meter
Made all needed changes to the Compiler so that we can build the program again. Ported FunctionCall::extract_stem and wrote preliminary logic for parsing function calls; most of that will need to be replaced when we finally have the code we need in order to write the code to identify the OpcodeFunction being invoked.

The script editor window now has a meter on the bottom which shows how much of the available MPVR space the game variant is using. However, the meter is currently 6 bytes short when testing with a resaved copy of Alpha Zombies. (We can see how many bytes are in use by looking at the bytecount encoded after MPVR's SHA-1 hash, but this bytecount doesn't include itself or the bytes that precede it (0x28 bytes in total).)