Compiler work: started on namespaces
Unique IDs for VariableScopeIndicatorValues are now integers, allowing us to use enums (which the compiler-as-in-MSVC can error-check for us). Wrote property definitions for all variable types. Modified the typeinfos for players and teams to specify that they have a non-zero static count (i.e. you can do typename[index] to refer to an absolute player/team).

Started work on namespace definitions.