Compiler documentation revisions
Made revisions to the compiler documentation for: abstract properties (the term is introduced and defined); aliases (to clarify circumstances in which multiple aliases can share a name); script traits (to clarify that we should handle those the same as widgets); additional syntax errors we need to catch in variable references; a full description of what information a resolved variable reference would consist of; a full description of how to identify property access in a variable reference; and a description of how to resolve an alias definition.

Wrote literal notes in the margins to make it explicit where to handle script traits and widgets, script stats, and script options.