Compiler work: resolving variable references
OpcodeArgTypeinfo now defines an internal name (used for script code, e.g. "player[0]"), a friendly name, and a description. All types have been amended to provide these.

DetailedEnum allows you to add a plain-signature info. The "which" enums (megalo_objects, megalo_players, and megalo_teams) use these so that compiler-related data can refer to them.

Started work on code to resolve variable references, as described in my notes. It's worth noting that doing this entirely from VariableReference (as opposed to using a member function on VariableReference::Part, as the JS implementation does) will cause complications with alias resolution when we get around to doing that.