Compiler work: variable reference work
Moved the variable reference code to its own file so I have some room to breathe while editing it.

I've decided that the best way to interpret variable references is to convert from "parts" (which can be namespaces, typenames, variables, members,...) to "interpreted parts" (each of which is ONE specific thing that is being refered to).