Compiler work: variable reference work
Overhauled how we resolve variable references. We now use a function to parse the first part(s) of the variable reference (so we can deal with namespaces and statically-indexable types like player), followed by a loop for the remaining pieces. This is more in line with how the JavaScript implementation behaved, and it gives us an opportunity to resolve aliases as well.

Our code resolves aliases before trying to match to built-ins; this allows us to avoid having to repeat the built-in match process in the event that an alias is being used. However, this requires us to be especially careful when accepting aliases. (An alternate approach would be to match aliases after trying everything else, and if the alias matches, decrement the loop index once so that the next loop iteration reprocesses the transcluded alias content.)