Compiler work: variable reference work
Variable reference is mostly done. There are two main tasks left to perform: implement VariableReference::_transclude, which is responsible for substituting an alias's contents in when we see the alias being invoked; and set up a flag so that attempts to access an abstract property fail if done from a function call argument list or a condition. The latter will be easier when we return to parsing statements.

Speaking of which: now that variable reference parsing is complete enough for external callers to use it, we can modify the logic for parsing assignments, calls, and comparisons to make use of it. We should begin doing that.