Compiler work: function call parsing work
Moved the argument array from Condition and Action to Opcode, to make things easier for the compiler.

The compiler's function-parsing code now actually retains compiled OpcodeArgValues. It does type-checking for return values, but still needs code to retain the return value.

Defined an OpcodeArgValue::compile overload which takes a Script::VariableReference; we can use this to easily compile variables.

The Assignment and Comparison parser classes now inherit from a Statement class which can hold an Opcode pointer. We still don't have code to actually store a compiled function call, but this will be the basis of that code.