Compiler work: function call work
Wrote up some short-term plans.

The Script::FunctionCall class is obsolete: function calls are now parsed the instant they're detected, rather than in multiple stages, so we don't need to retain any information about them. This also makes the Script::Assignment class obsolete, since the only difference between it and a Script::Comparison was the ability to store a FunctionCall* as the righthand side. As such, Script::Assignment has been absorbed into Script::Statement (and Comparison soon will be, too), and Script::FunctionCall has been wholly removed.

The code for parsing function calls now creates an opcode of the proper type (Condition or Action; we should never create bare Opcodes), and the opcode is now retained post-parsing when the final Statement is committed.