Compiler work
Made it possible for the compiler to log errors at arbitrary positions. Function call parsing now uses this to log errors with the line and col numbers of the start of the function's context/name, even if we can only detect these errors later in parsing.

Declared Compiler::_applyConditionModifiers, which will handle the "not", "and", and "or" keywords. The function has not yet been coded; we can't do that until we have code to compile comparisons. It is called from the code to parse functions.