Compiler work
The code for compiling assignments now checks for unsupported operators in assignments to abstract properties. The code for compiling function calls now checks to see if we are attempting to assign the result of a function call to a read-only value or abstract property.

Corrected the presence and absence of the "read-only" flag on some VariableScopeIndicatorValues.