Compiler to-do list update
Came up with a specific plan for preventing the decompiler from mishandling user-defined functions and choking on recursive triggers.

We need to finish memory management for most of the compiler constructs.

Came up with a specific plan for tracking in-scope aliases and user-defined functions.

When the compiler is functional and round-trip decompile/recompile tests confirm it works properly, we need to test to see if user-defined functions actually work in-game; right now, they're a theoretical concept: the file format should make them possible but it's unknown whether the script engine can cope with them.