Compiler work: user-defined functions
Compiler now has a destructor and destroys any compiler state objects it still has a hold of.

User-defined functions have Triggers created for them immediately upon their opening (in contrast to how other Blocks will work), so that we can properly compile calls to them as those calls are encountered. The Compiler tracks what user-defined functions are currently in-scope and remembers their trigger indices.

Calls to user-defined functions will throw a parse error if any arguments were passed, or if the terminating ")" for arguments was omitted. Corrected the error message for calls to user-defined functions that appear inside of a block's conditions. Calls to user-defined functions are now fully compiled.

Fixed the removal of out-of-scope aliases: we now check alias->parent instead of alias->owner.