Compiler work: assignment work
Realized that we have a problem with property/accessor interactions: "score" is a property, but there is also a setter-accessor for it. (KSoft.Tool doesn't have the property listed as read-only, but maybe it should.) This presents a problem: our code can't cope with a read-only property that has a setter accessor.

Started work on compiling assignment statements. Standard assignments can fully compile; accessors are in progress.