Compiler work: conditions
Fixed compile errors.

The code to parse comparisons is now complete, including the not/and/or keywords. I had to reintroduce the Comparison class because the and/or linkage for conditions is actually relative to the containing Trigger and so can only be resolved when the containing Block(s) are compiled.

VariableReference now fails if its first part resolves to an absolute alias of an imported word.

Wrote OpcodeArgValueVector3::compile, just to have it on hand as a basic example of how to override that OpcodeArgValue method.

I examined Freeze Tag's loaded content and wrote down how conditions' or-group values should be set up.