Compiler to-do list
Decided on an approach for handling the "score" property/accessor.

Decided on the overall approach for writing code to compile variables.

Realized a problem with how OpcodeArgValue::compile works: it will not be possible to compile string references with the current system. Wrote down some detailed plans for remedying that.