Compiler work: function error handling
Function parsing no longer uses exceptions.

User-defined function declarations with invalid names will no longer throw exceptions; all cases are either fatal or non-fatal errors.

If you attempt to assign a string literal to a variable, or write a comparison statement with a string literal on the righthand side, the parser will attempt to skip the string literal after logging the error. If the parser is unable to find the closing quotation mark, then parsing will fail with a fatal error.

If the "on" keyword runs into a non-fatal error, and fails to skip to the next ":" glyph, then parsing will fail with a fatal error.