Compiler work: arg_compile_result changes
Changed (arg_compile_result) from an enum to a struct, allowing OpcodeArgValue::compile to send detailed error messaging and additional information back to the Compiler. All types that had (compile) overrides defined have been updated.

One of the struct's new features is the ability to signal whether the OpcodeArgValue needs more script arguments, separately from signalling failure, and the ability to signal whether a failure is irresolvable. This makes it possible for us to continue parsing arguments to a function call even if one of them is invalid, outside of specific cases where the invalid argument was needed to understand subsequent valid ones. For example, an invalid shape type would be an irresolvable failure that prevents us from parsing further arguments (because the shape type indicates how many more arguments there are), whereas an invalid coordinate in a Vector3 would not prevent us from parsing the remaining Vector3 coordinates or other arguments.