Compiler to-do list
Wrote some plans for how we could validate script access to the "symmetry" setter and to the "death_event_damage_type" getter.

Wrote some plans for how we could make the "death event damage type" enum accessible to script authors, so that they don't have to use bare integers or make their own aliases.

Removed some to-do list items that are complete and that I forgot to remove in the last commit.

Added the (de)compiler UI needs to the to-do list.