Compiler work: Variable declaration work
The compiler emits an error if you try to declare a property or accessor, or if you try to use a property or accessor as the initial value in a variable declaration.

Wrote the compiler logic for actually generating variable declaration data after reading a declaration in script.

When parsing variable declarations, the compiler now uses unique_ptrs for the variable being declared and for its initial value.

Redesigned how variable declarations are retained in memory: they now use a single, generic class, which means less special-casing for the compiler.