Redesigned the variable declaration syntax
Redesigned the variable declaration syntax.

Redeclaring a variable is an error if the redeclaration provides a network type or initial value when one was already provided in a previous declaration.

Declaring a variable whose initial value is another variable is an error.

If a declared variable's type cannot have a networking mode or an initial value, providing one is now an error instead of a warning.

Initializing a variable to a floating-point number yields a specific error message.

Declaring a nested variable (e.g. global.player[0].number[0] as opposed to player.number[0]) is an error.