Compiler work: implicit variable declarations
The Compiler now creates VariableReferences exclusively through a helper function (except when parsing variable declarations); this helper function allows us to create implicit variable declarations whenever a script variable is used. Probably for the best that we went and did this because it looks like we'd forgotten to call VariableReference::resolve in a few places. Yikes! The helper function does that for us now.

When a VariableDeclaration is created for a type with a numeric initial value, it now sets its OpcodeArgValueScalar to constant zero.

Changed how OpcodeArgValueScalar::is_const_zero works, taking advantage of some earlier changes to how variables store their values.