Compiler work: compiling variables
Wrote the code to compile each variable type (but not yet the "any variable" or "player or group" types). The majority of cases are handled by the base Variable class, thanks to some extensions to the scope-enum-list definition objects. Some cases are specific to each subclass and can only be handled there; specifically: constant integers (OpcodeArgValueScalar) and the "biped" property (OpcodeArgValueObject).