Compiler work: critical fixes
Fixed a critical error in the code to resolve widget arguments; the error involved the use of (auto) instead of (auto&) leading to us copying the game variant data in full and then destroying the copy (which trashed pointers in the original).

We now only try to generate implicit variable declarations for VariableReferences that actually refer to variables (as opposed to, say, script_widget[0]).

Improved the code to handle the case of a function call to a function with no arguments.

Fixed a bug: when a function's arguments failed to parse, we accidentally skipped to the next ")" AFTER the affected function call, due to a mistake made when writing error-handling code.

Fixed string_scanner::extract_up_to_any_of to skip to and past the matching character, if any.

Fixed a bug: passing only a const char* to arg_compile_result::failure would result in the string literal implicitly casting to a boolean instead of to a QString (because C++ is a cursed language). We now have an overload on arg_compile_result specifically for const char*.

The "const bool" type now logs a better error message if the argument received is blank. (Wondering whether I shouldn't handle that in the compiler itself, though.)

Script stats are no longer flagged as read-only (because you can write to their per-player and per-team run-time values).