Compiler work: line number fix
A line break at the start of the script will no longer cause line numbers in all logged warnings and errors to be off by one.

Came up with specific plans for fully implementing else(if)-blocks, as well as a plan for testing to ensure that our compiler matches Bungie's behavior without having to bring the decompiler in line first.