Compiler fixes
Fixed major issues in cobb::indexed_list and cobb::pointer list.

Fixed the compiling of widget meter parameters.

Fixed the loading of icon values.

Fixed the compiling of for-each-object-with-label values. Discovered that a label is required; the "no label" syntax is planned to be removed.

The compiler can now compile an empty string, erasing all script content in the variant.