Minor compiler fixes and UI improvements
Removed the "for each object with no label" syntax, since MCC doesn't consider that valid and will assume that a file using it is corrupt. Added an end-of-compile check for for-each-object-with-label loops that don't specify a label.

Added notes to some Megalo actions based on in-game testing. Currently these descriptions are not displayed in the program.

Fixed the opcode-to-script mapping for the "player.add_weapon" action.

The variant space usage meter should now react to the variant being recompiled. The meter should now remain at its minimum size, instead of stretching whenever the script editor's main "content" pane doesn't use the full height of the window.

When deleting a string from the string table, the UI should now auto-select the next string (if there is one). Modified other lists in the script editor to behave similarly (previously, they selected the list item before the one you deleted; selecting the one after feels more natural).

Wrote down some miscellaneous observations in a text file until I have a place inside the program to put them.