Object type testing
Identified several object types that seem to be defined in the upcoming Thorage update. They can be spawned by Megalo on the current release build of MCC Reach.

Fixed a crash that could occur when emptying out one of the UI lists (e.g. strings, widgets, stats, etc.).

The compiler now allows you to use integer aliases to specify a Forge label index for a for-each-object-with-label loop.