Compiler fixes
(De)compiling of string literals now properly generates and decodes escape sequences; strings that contain "\r\n" should no longer be handled inconsistently by the compiler and decompiler, and escape codes should now be supported when compiling.

The string "--" should no longer be misinterpreted as the start of a line comment if it occurs inside of a string literal.

Updated some actions based on information that was sent to me, based on information visible in Race, and based on tests I performed. Identified the boolean parameter for "object.attach_to". Made "player.try_get_weapon" and "player.get_weapon" use an enum instead of a bool to specify the weapon slot.

Fixed the compiling of "or"-linked conditions.

Fixed the compiling of nested variables wherein the top-level and nested variables were both of the same type, and that type used "which" rather than an index for its top-level values (e.g. current_player.player[0]).

Fixed a bug in the script editor layout which would cause the window to have a massive minimum size if a string in the string table had an especially long length.