UI improvements for string management
Compiler: constant integers now warn if they can't fit in a 16-bit signed value.

When you create a new Forge label, the UI now selects it.

The string table editor now has a button to copy a string's text, code, or index.

The unresolved string reference window will now show the number of strings (by way of a suffix in the number-select), and the number of references to the currently-selected string.