Trigger decompiling fix
Fixed a check in TriggerDecompileState::check_if_is_function which accidentally looked at the number of callees rather than the number of callers.

Ran several tests on object attachment and vehicles, and documented the results. Tested and confirmed that user-defined functions work when called from multiple places, including when called recursively: there doesn't appear to be any risk of a double-free crash when the script is unloaded (as would be the case if the runtime used pointers to maintain nested triggers and gave the nesting triggers ownership of the nested triggers).