Redesigned the code editor UI
The script editor now uses a QTextEdit instead of a QPlainTextEdit, as preparation for implementing syntax highlighting.

Compiler warnings and errors are now shown in a list pane. Double-clicking on a list item jumps to its line in the code. A custom QListWidget subclass allows us not only to let the user copy multi-selected log items, but also to transform those log items before copying (allowing us to insert the line and column numbers into the copied output even when they aren't displayed in the list editor directly). The list pane can be filtered to show only warnings or only errors, by way of UI controls that mimic the style of most browsers' JavaScript consoles.

When the compiler shows error messages for bad Forge label names, the error messages now include the label name as a string literal. (This change was made to both the compiler code, for for-each-object-with-label loops, and to the Forge label opcode argument type.)