Compiler fixes
Aliases of negative constant integers will now be parsed properly.

Fixed a bug with nested functions: the compiler accidentally inserted a call to these functions into the containing block when compiling.

When the compiler reports that a script argument failed to parse, it now reports argument indices starting at 1, which is more intuitive. The compiler also now properly reports error messages for calls to functions in the "game" namespace.

The Required Object Types UI now allows you to see an object type's description in a tooltip.