Localized string library work
Documented the build environment for this program in the readme. I've been meaning to do this for a while, but never got around to digging up the links; I ended up stumbling across them while wasting two hours debugging Qt resource shenanigans.

Moved the Qt resource file.

Fixed memory management for the localized string library. The localized string editor's handling of the library UI should be done; the library UI itself is not, and does not modify the string. (It signals that a modification occurred, so any user changes to the textboxes in the localized string editor will be blown away, but this is more akin to a UI reset than to actually applying changes.)

The string library UI now properly generates a listview with all hardcoded strings. The buttons to preview different languages work properly. Filtering the list view doesn't work; applying changes doesn't work; we don't have any code for displaying tokens and allowing the user to edit them (e.g. letting the user insert any number into "X points").