Engine tests and UI tweaks
We now use the text "Spanish (Latin America)" to identify the language that was listed in our source material as "Mexican."

Added some strings from Invasion Slayer to the localization library.

Tested for a maximum call depth. The engine doesn't seem to have one.

Tested detaching a player's carried weapons and arrmor abilities. Recorded the results.