String table now caches savedata
ReachStringTable now writes to a private cobb::bitwriter, and uses this to retain whatever output data it generates. This output data will be reused if the table and its contents are not modified. This allows the script space usage meter in our UI to more efficiently measure the string table's space usage.

Renamed a few member functions on ReachString.