Fixed compiler CTD (non-member call with bad return value)
Fixed a crash that could occur if a non-member function with a return value was called and assigned to a variable of the wrong type (or an indeterminate type, as in the case of a bad alias on the lefthand side).

Tested several ideas in-game and wrote down the results.