Improved error messaging
We now have a specific error message for a statement which begins with an operator. This should help in cases where someone forgets that expressions aren't supported.

Wrote down some plans for how we could avoid crashing or saving a corrupt file if we're asked to save a game variant whose content can't fit in the spaces allotted by the file format.