isBuiltIn dealt with; started on save failsafes
We now forcibly clear GameVariantDataMultiplayer::isBuiltIn on load.

Started work on systems to allow you to always save a game variant, even if the file format's limits are exceeded. We'll basically make this work by stuffing overflowing data into a custom file block, though UI messaging will need to make it VERY clear to the user that we've done this, and that resaving the file through any other tools will cause this custom file block to be lost.