Work on save failsafes
Defined ReachStringTable::write_placeholder and ReachStringTable::write_fallback_data. Additionally, ReachStringTable::write now returns a success bool. The two methods that interact with bitwriters will not advance the bitwriter in the event of a failure (write_placeholder will rewind the bitwriter if a problem is encountered, while write never even touches the bitwriter if we fail to set up valid data).