Save process
Implemented a "save process" object instead of having game variants' top-level data structures write to a bare stream. This allows game variants to signal details about what was saved -- in particular, whether certain parts of the file had to be written to the xRVT block.

Using this, ReachVariantTool now shows an additional confirmation prompt (with an explanation) if the file you're about to save would be incomplete without an xRVT block.

An additional improvement to the file save process is the use of QSaveFile instead of QFile.

Finally, the classes for loaded xRVT data have been moved to their own file, so hopefully we can adjust those without having to recompile 80% of the program now.