Prep for string management and help docs work
Cleaned up the to-do list, and consolidated the string management tasks.

Moved debug code out of the main window file and into its own file. This'll make it easier to add code for generating help documentation XML without completely making a mess of the main window file.

Added member functions to GameVariantDataMultiplayer with the intention of making it easy to check whether a string is in use by particular elements of a game variant (i.e. which traits, which option, which Forge label). This should come in handy when tackling some of the string management tasks.