String management and UI improvements
Editing the name or description of a scripted player trait set through a means other than the player trait UI (e.g. the string table UI) will now send the appropriate ReachEditorState signal indicating that the trait set has been modified. Ditto for script options and their values. ReachEditorState is responsible for detecting changes to the content of a string that is in use by an option or trait set; the UIs for options and traits still need to handle switching options and traits from one string to another.

If a string is in use as a Forge label, then editing it through any means (even outside of the Forge label UI) will constrain you to editing it as a single-language string. This is handled within the localized string editor dialog's code.

Added ReachStringPicker::selectedStringSwitched, to facilitate the above changes.

When creating a new script option or script option value, the newly-created entity is now properly selected in its respective list widget.

The script options UI now reacts to external string changes in a more efficient manner i.e. it only refreshes the option list if the string belonged to any option, and it only refreshes the value list if the string belonged to the currently selected option.

ReachStringPicker::clearTarget now visually clears the string-picker widget.