Plans
Plans to rename the current "else" and "elseif" statements to "alt" and "altif", since they do not fully meet the requirements of an "else" statement (i.e. statefulness).

Wrote down an idea for how to implement true else and elseif statements, though I won't be pursuing that idea.

Started work on XML-based documentation, which I can either figure out how to display in-app or convert to HTML. Not sure if I want to finish this manually, or auto-generate it and then flesh it out. Might just take the former approach, to avoid writing code that'll only be useful once.