Renamed "else(if)" to "alt(if)"
Our implementation of "elseif" blocks was an imperfect approximation of elseif blocks in other languages, and was missing some key behaviors. It would take a fair amount of effort and overhead to implement true elseif blocks (see relevant file in the "docs" folder). For now, the current block has been renamed to "alt" and "altif" i.e. "alternatively". The compiler reserves "else" and "elseif" for future use.