User-defined enums are now implemented
Implemented user-defined enums.

Added a predefined enum for the values returned by the object.get_orientation action.

Fixed an oversight that would've allowed you to shadow the names of namespaces using some constructs, including user-defined functions. This stemmed from a missing check in the "check if a name is taken" function on Compiler.