Opcode arg type: fireteam list
Turns out, the object.set_spawn_location_fireteam action doesn't take a fireteam index, but rather a bitmask listing multiple fireteam indices. I've changed the parameter from type ConstSInt8 to a new type which accepts the words "none", "all", or a comma-separated list of fireteam indices.

Renamed the action, as well, to object.set_spawn_location_fireteams. Note the plural.

I am continuing to decompile and document the official gametypes. The only (important) ones that remain are Infection and Invasion, the latter of which I'm currently working on.