Fixed crash when deleting an option's strings
Fixed a crash that could occur from the following steps: select a script option and a script option value; delete the option; and then go to the string table editor and delete the strings for the option and its values. The crash occurred because deleting an option wouldn't properly sever the link between the option-value-editing UI and the string references in the selected option value.

Did some more work on script API documentation.