UI fixes and improvements
Added debug information to ReachStringPicker to make it easier to investigate crashes related to it.

Moved a lot of data fix-up code from the option-editing UI to member functions on the options.

Deleting a Forge label now selects the next Forge label in the list (or, if there is none, the previous). Ditto for option values. This makes these elements consistent with other lists-of-things that you can edit through the UI.

WIP help docs: added object.enable_spawn_zone and, after quite a bit of testing, object.set_garbage_collection_disabled. Corrected the opcode definition for the latter as well.