"HUD" member fixes; documentation updates
Fixed namespace and "which" definitions for the HUD player team and the HUD target player team.

In-game testing failed to determine the precise function of unk_14_team and unk_15_team; in fact, I was unable to get the UI to display values for anything other than member variables on the HUD player team.

Updated several MOTL definitions after in-game testing; documentation has received matching updates along with additional corrections.

Documentation work: finished the listing for the damage reporting types; files for the main window and script window now link to files for each window page; documentation for the "for each team" loop now indicates that it also runs for Neutral Team.

The script editor is now available when the program is compiled in Release configuration. I'm dealing with some tech trouble at the moment and I want to make sure I can still ship this if my computer stops being usable.