Documentation build system work
Started work on code to load the full XML for a script type or namespace, so we can process it as needed (e.g. making <related/> tags bidirectional) and then write out all needed pages.

Removed an unused field from cobb::xml::token. That field was probably a leftover I had planned for when working on NorthernUI, an Oblivion mod.