Fixes; documentation build system work
Fixed a mistake in cobb::sprintf that could cause the last character of the format string to be lost, if the result string was too long.

Fixed an XML syntax error in object.xml. Filled in missing details in a couple other XML files. Adjusted the CSS for the object type list.

We now have classes for loading API documentation data and writing HTML output. The load process is not complete, however, and neither is the HTML write process. Currently we can generate the base pages for types, but not pages for their members; we also don't load properties or accessors at all yet.