Documentation build system work
Relationships between member functions of the same class will now be bidirectional.

Code for loading relationships has been moved from APIMethod to APIEntry, which will allow us to reuse it for properties and accessors.

We now properly load notes and write them to the page.

The pages now have JavaScript code which will offer a permalink to any permalinkable subheading.

Improved the description of all vector arguments in object.xml: it is now made explicit that you must supply an integer constant and not, say, a number variable.